requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
insightface==0.7.3
onnxruntime==1.16.3
opencv-python==4.8.1.78
numpy==1.24.4
pillow==10.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
jinja2==3.1.2
aiofiles==23.2.1
websockets==12.0
imutils==0.5.4
scikit-learn==1.3.2 